tests/test_model.py
import unittest
from model.train import train_model
from model.model import create_model, load_model_from_file

class ModelTestCase(unittest.TestCase):
    def test_create_model(self):
        """Test the creation of the model."""
        model = create_model()
        self.assertIsNotNone(model)
        self.assertTrue(hasattr(model, 'predict'))

    def test_train_model(self):
        """Test if the training function completes without errors."""
        try:
            train_model()
        except Exception as e:
            self.fail(f"train_model() raised an exception: {e}")

    def test_load_model_from_file(self):
        """Test loading a model from a file."""
        model = load_model_from_file('path/to/saved_model.h5')
        self.assertIsNotNone(model)
        self.assertTrue(hasattr(model, 'predict'))

if __name__ == '__main__':
    unittest.main()
